refactor(channels): consolidate the flat streaming fallback into the deprecation module
